feat(search): implement all Similarity models for Phase 50 (GC-859 to GC-867)
- Added AxiomaticSimilarity: axiomatic retrieval model with s and k parameters
- Completed BM25Similarity: BM25SimWeight and BM25SimScorer with full formula
- Added DFRSimilarity: Divergence From Randomness with BasicModel, AfterEffect, Normalization
- Added IBSimilarity: Information-Based model with Distribution, NormalizationIB
- Added IndriSimilarity: Indri retrieval model with Dirichlet smoothing
- Added LMDirichletSimilarity: Language model with mu parameter
- Added LMJelinekMercerSimilarity: Language model with lambda interpolation
- Added MultiSimilarity: Combines multiple similarities with weighted sum
- Updated Similarity interface: Added SimWeight interface and Coord method
- Fixed ClassicSimilarity: Changed Coord/QueryNorm return types to float32
- Updated all implementations for interface compliance

Closes: GC-859, GC-860, GC-861, GC-862, GC-863, GC-864, GC-865, GC-866, GC-867

+++ search/classic_similarity_test.go
@@ -96,15 +96,15 @@ func TestClassicSimilarity_QueryNorm(t *testing.T) {
 
 	// QueryNorm(4) = 1/sqrt(4) = 0.5
 	result := sim.QueryNorm(4.0)
-	expected := 0.5
-	if math.Abs(result-expected) > 0.0001 {
+	expected := float32(0.5)
+	if math.Abs(float64(result-expected)) > 0.0001 {
 		t.Errorf("QueryNorm(4.0) = %f, expected %f", result, expected)
 	}
 
 	// QueryNorm(1) = 1/sqrt(1) = 1.0
 	result = sim.QueryNorm(1.0)
-	expected = 1.0
-	if math.Abs(result-expected) > 0.0001 {
+	expected = float32(1.0)
+	if math.Abs(float64(result-expected)) > 0.0001 {
 		t.Errorf("QueryNorm(1.0) = %f, expected %f", result, expected)
 	}
 }
@@ -115,17 +115,17 @@ func TestClassicSimilarity_Coord(t *testing.T) {
 	tests := []struct {
 		overlap    int
 		maxOverlap int
-		expected   float64
+		expected   float32
 	}{
 		{1, 1, 1.0},
-		{2, 3, 2.0 / 3.0},
+		{2, 3, float32(2.0 / 3.0)},
 		{3, 3, 1.0},
 		{0, 5, 0.0},
 	}
 
 	for _, test := range tests {
 		result := sim.Coord(test.overlap, test.maxOverlap)
-		if math.Abs(result-test.expected) > 0.0001 {
+		if math.Abs(float64(result-test.expected)) > 0.0001 {
 			t.Errorf("Coord(%d, %d) = %f, expected %f", test.overlap, test.maxOverlap, result, test.expected)
 		}
 	}
